refactor(data-model): the native-object predicate joins the other type checks
`isValidFabricNativeObject()` asks a membership question and belongs with the
rest of them. It could not live there while the only way to answer it was
through a module holding the fabric classes; after #6503 took the tag
vocabulary and the class read out into files that import nothing heavy, it can.

What is left in `native-builtin-tags.ts` is `tagFromNativeBuiltinClass()`, which
both `type-check.ts` and `native-type-tags.ts` now ask -- the shared leaf the
split existed for, with the part that was only ever borrowing it moved out.

Its tests follow it into `type-check.test.ts`.
