+++ packages/data-model/src/fabric-value.ts
@@ -59,6 +59,6 @@ export {
   shallowFabricFromNativeValue,
 } from "./native-conversion.ts";
 
-export { isValidFabricNativeObject } from "./native-builtin-tags.ts";
+export { isValidFabricNativeObject } from "./type-check.ts";
 
 export { valueEqual } from "./valueEqual.ts";
